Updated premium.html
I update the popup menu that users get when they click purchase. I did this because some people on the support server said it was confusing which is why they did not join the support server before buying premium. The old message was:
`<p>
                        Before proceeeding to payment, please join our
                        <a href="/support" target="_blank" class="link-dark">support server</a> to receive the role and
                        rewards.
                    </p>`

and the new message is:
`<p>
                        Please join our <a href="/support" target="_blank" class="link-dark"> support server</a> <strong>BEFORE</strong> purchasing so that your premiun can be successfully enabled.
                    </p>
                    <p>
                        <i>You will not receive your perks propelry if you do not join the supposrt server.</i>
                    </p>`