Clarify reason to login
Summary--
Clarifies why users should login when they first encounter the login prompt in DMs. The hope is to alleviate some of the confusion since this is a non-standard thing for many bots and streamline the process by adding a brief description of why it is necessary by mentioning that this is required for the bot to function and providing a simple explanation.