Issues #101, #103
__tools.py__
Reduced the number of servers required for the server list from **10** to **9** to be more aesthetically pleasing (or at least I hope the change I made did this). Removed the number 10 reaction at the same time

Also added a note saying that users with fewer than 10 servers will not be able to flip pages.

__core.py__
Added a reason to the `blacklist` command (and attempted a db addition) and added a reason section to the `viewblacklist` command

__snippet.py__
Added a counter to the `viewsnippet` command to show how many snippets have been added