Update core.py
changed the whitelist command to do a check against `discord.user` instead of the custom `MemberConverter`. This is because if a member is not in the server then the bot can not resolve them to have them removed from the blacklist.

This could be superseded by a `USerConverter` or a check in the `MemberConveter to check if the user is fectable.